@@ -3,68 +3,67 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace RedditBots.Framework
+namespace RedditBots.Framework;
+
+/// <summary>
+/// Base class for implementing a long running <see cref="IHostedService"/>.
+/// </summary>
+public abstract class BackgroundService : IHostedService, IDisposable
 {
+    private readonly CancellationTokenSource _stoppingCts = new CancellationTokenSource();
+
     /// <summary>
-    /// Base class for implementing a long running <see cref="IHostedService"/>.
+    /// The task returned from ExecuteAsync.
     /// </summary>
-    public abstract class BackgroundService : IHostedService, IDisposable
+    public Task ExecutingTask { get; private set; }
+
+    /// <summary>
+    /// This method is called when the <see cref="IHostedService"/> starts. The implementation should return a task that represents
+    /// the lifetime of the long running operation(s) being performed.
+    /// </summary>
+    /// <param name="stoppingToken">Triggered when <see cref="IHostedService.StopAsync(CancellationToken)"/> is called.</param>
+    /// <returns>A <see cref="Task"/> that represents the long running operations.</returns>
+    protected abstract Task ExecuteAsync(CancellationToken stoppingToken);
+
+    /// <summary>
+    /// Triggered when the application host is ready to start the service.
+    /// </summary>
+    /// <param name="cancellationToken">Indicates that the start process has been aborted.</param>
+    public virtual Task StartAsync(CancellationToken cancellationToken)
     {
-        private readonly CancellationTokenSource _stoppingCts = new CancellationTokenSource();
+        // Store the task we're executing
+        ExecutingTask = Task.Run(() => ExecuteAsync(_stoppingCts.Token), cancellationToken);
 
-        /// <summary>
-        /// The task returned from ExecuteAsync.
-        /// </summary>
-        public Task ExecutingTask { get; private set; }
+        return Task.CompletedTask;
+    }
 
-        /// <summary>
-        /// This method is called when the <see cref="IHostedService"/> starts. The implementation should return a task that represents
-        /// the lifetime of the long running operation(s) being performed.
-        /// </summary>
-        /// <param name="stoppingToken">Triggered when <see cref="IHostedService.StopAsync(CancellationToken)"/> is called.</param>
-        /// <returns>A <see cref="Task"/> that represents the long running operations.</returns>
-        protected abstract Task ExecuteAsync(CancellationToken stoppingToken);
-
-        /// <summary>
-        /// Triggered when the application host is ready to start the service.
-        /// </summary>
-        /// <param name="cancellationToken">Indicates that the start process has been aborted.</param>
-        public virtual Task StartAsync(CancellationToken cancellationToken)
+    /// <summary>
+    /// Triggered when the application host is performing a graceful shutdown.
+    /// </summary>
+    /// <param name="cancellationToken">Indicates that the shutdown process should no longer be graceful.</param>
+    public virtual async Task StopAsync(CancellationToken cancellationToken)
+    {
+        // Stop called without start
+        if (ExecutingTask == null)
         {
-            // Store the task we're executing
-            ExecutingTask = Task.Run(() => ExecuteAsync(_stoppingCts.Token), cancellationToken);
-
-            return Task.CompletedTask;
+            return;
         }
 
-        /// <summary>
-        /// Triggered when the application host is performing a graceful shutdown.
-        /// </summary>
-        /// <param name="cancellationToken">Indicates that the shutdown process should no longer be graceful.</param>
-        public virtual async Task StopAsync(CancellationToken cancellationToken)
+        try
         {
-            // Stop called without start
-            if (ExecutingTask == null)
-            {
-                return;
-            }
-
-            try
-            {
-                // Signal cancellation to the executing method
-                _stoppingCts.Cancel();
-            }
-            finally
-            {
-                // Wait until the task completes or the stop token triggers
-                await Task.WhenAny(ExecutingTask, Task.Delay(Timeout.Infinite, cancellationToken));
-            }
-
-        }
-
-        public virtual void Dispose()
-        {
+            // Signal cancellation to the executing method
             _stoppingCts.Cancel();
         }
+        finally
+        {
+            // Wait until the task completes or the stop token triggers
+            await Task.WhenAny(ExecutingTask, Task.Delay(Timeout.Infinite, cancellationToken));
+        }
+
+    }
+
+    public virtual void Dispose()
+    {
+        _stoppingCts.Cancel();
     }
 }
