@@ -1,8 +1,7 @@
-﻿namespace RedditBots.PapiamentoBot.Models
+﻿namespace RedditBots.PapiamentoBot.Models;
+
+public class Request
 {
-    public class Request
-    {
-        public string Content { get; set; }
-        public string From { get; set; }
-    }
+    public string Content { get; set; }
+    public string From { get; set; }
 }
