@@ -1,18 +1,13 @@
 ﻿using RedditBots.PapiamentoBot.Settings;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace RedditBots.PapiamentoBot.Models
+namespace RedditBots.PapiamentoBot.Models;
+
+public class Response
 {
-    public class Response
-    {
-        public static Response Empty() => new Response();
+    public static Response Empty() => new();
 
-        public bool MistakeFound { get; set; }
+    public bool MistakeFound { get; set; }
 
-        public Word Mistake { get; set; }
-    }
+    public Word Mistake { get; set; }
+
 }
