@@ -1,24 +1,24 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace RedditBots.PeriodicallyBot.Settings
+namespace RedditBots.PeriodicallyBot.Settings;
+
+public class PeriodicallyBotSettings
 {
-    public class PeriodicallyBotSettings
-    {
-        public TimeSpan TimeOfExecution { get; set; } = TimeSpan.FromHours(13);
+    public TimeSpan TimeOfExecution { get; set; } = TimeSpan.FromHours(13);
 
-        public List<PeriodicTask> PeriodicTasks { get; set; } = new List<PeriodicTask>();
-    }
-
-    public class PeriodicTask
-    {
-        public int DayOfTheMonth { get; set; }
-
-        public TaskType TaskType { get; set; }
-    }
-
-    public enum TaskType
-    {
-        PostToCSharpMonthlyThread
-    }
+    public List<PeriodicTask> PeriodicTasks { get; set; } = new List<PeriodicTask>();
 }
+
+public class PeriodicTask
+{
+    public int DayOfTheMonth { get; set; }
+
+    public TaskType TaskType { get; set; }
+}
+
+public enum TaskType
+{
+    PostToCSharpMonthlyThread
+}
+
