@@ -1,9 +1,8 @@
 ﻿using System.Collections.Concurrent;
 
-namespace RedditBots.Libraries.Logging
+namespace RedditBots.Libraries.Logging;
+
+public class HttpLoggerQueue
 {
-    public class HttpLoggerQueue
-    {
-        public readonly ConcurrentQueue<HttpLogEntry> Messages = new ConcurrentQueue<HttpLogEntry>();
-    }
+    public readonly ConcurrentQueue<HttpLogEntry> Messages = new ConcurrentQueue<HttpLogEntry>();
 }
