@@ -1,27 +1,25 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using RedditBots.Web.Helpers;
 using RedditBots.Web.Models;
-using System.Threading.Tasks;
 
-namespace RedditBots.Web.Controllers
+namespace RedditBots.Web.Controllers;
+
+[ApiController, Route("[controller]/[action]")]
+[ApiKeyRequired]
+public class ApiController : ControllerBase
 {
-    [ApiController, Route("[controller]/[action]")]
-    [ApiKeyRequired]
-    public class ApiController : ControllerBase
+    private readonly LogHandler _logHandler;
+
+    public ApiController(LogHandler logsHelper)
     {
-        private readonly LogHandler _logHandler;
+        _logHandler = logsHelper;
+    }
 
-        public ApiController(LogHandler logsHelper)
-        {
-            _logHandler = logsHelper;
-        }
+    [HttpPost]
+    public async Task<IActionResult> Log([FromBody] LogEntry entry)
+    {
+        await _logHandler.LogAsync(entry);
 
-        [HttpPost]
-        public async Task<IActionResult> Log([FromBody] LogEntry entry)
-        {
-            await _logHandler.LogAsync(entry);
-
-            return Ok();
-        }
+        return Ok();
     }
 }
