@@ -1,13 +1,12 @@
-﻿namespace RedditBots.Web.Models
+﻿namespace RedditBots.Web.Models;
+
+public class LogEntry
 {
-    public class LogEntry
-    {
-        public string LogName { get; set; }
+    public string LogName { get; set; }
 
-        public string Message { get; set; }
+    public string Message { get; set; }
 
-        public string LogLevel { get; set; }
+    public string LogLevel { get; set; }
 
-        public bool Notify { get; set; } = true;
-    }
+    public bool Notify { get; set; } = true;
 }
